Move the Date provider to a second file
